Bring GIF support to latest! Branch is useless now, I'll delete after this push.
It has a GUI option because it uses modulebase now, but idk what to put in there. I guess I could switch it from using commands to having a multitoggle of all the gifs in your folder or something?

Anyway, apparently the branch's code is actually really good and accounts to all those edge cases i was worrying about earlier.